Fix formatting for datetime widget
