+++ packages/material-ui/src/DateTimeWidget/DateTimeWidget.tsx
@@ -1,7 +1,15 @@
 import React from "react";
+import { utils } from "@rjsf/core";
 import TextWidget, { TextWidgetProps } from "../TextWidget";
 
+const { localToUTC, utcToLocal } = utils;
+
 const DateTimeWidget = (props: TextWidgetProps) => {
+  const value = utcToLocal(props.value);
+  const onChange = (value: any) => {
+    props.onChange(localToUTC(value));
+  };
+
   return (
     <TextWidget
       type="datetime-local"
@@ -9,6 +17,8 @@ const DateTimeWidget = (props: TextWidgetProps) => {
         shrink: true,
       }}
       {...props}
+      value={value}
+      onChange={onChange}
     />
   );
 };
